Airflow/requirements.txt
numpy
pandas
apache-airflow
kafka-python
